feat(user): add test endpoint for updating multiple users
Add a new GET endpoint `/test` to the User route to test the functionality of updating multiple users. The `updateManyUser` method in the UserService has been modified to update users with `isPayment` set to false, changing it to true. This change ensures that the endpoint can be used to verify the bulk update functionality.